[HWAN] Fixed the CityController and some minor fixes in other classes

+++ src/main/java/sample/data/jpa/web/CityController.java
@@ -1,12 +1,15 @@
 package sample.data.jpa.web;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import sample.data.jpa.domain.City;
 import sample.data.jpa.service.CityDao;
+
+import java.util.Objects;
 
 @Controller
 @RequestMapping("/city")
@@ -16,35 +19,42 @@ public class CityController {
     private CityDao cityDao;
 
     @PostMapping(value = "/create")
-    public ResponseEntity create(@RequestBody City city) {
-
-        System.out.println(city.getCountry());
-        String cityId = "";
+    public ResponseEntity<?> create(@RequestBody City city) {
+        if (Objects.nonNull(city)) {
             cityDao.save(city);
-            cityId = String.valueOf(city.getId());
-
-        return ResponseEntity.ok(city);
-    }
-
-    @RequestMapping("/delete")
-    @ResponseBody
-    public String delete(Long id) {
-        try {
-            cityDao.deleteCityById(id);
-        } catch (Exception ex) {
-            return "Error deleting the city: " + ex;
+            // Add header to return id in headers.
+            // If we remove the @JsonIgnore annotation over id in PO, then the id will be contained in returned object
+            // Therefore this fragment could be considered useless
+            var headers = new HttpHeaders();
+            headers.add(HttpHeaders.ACCESS_CONTROL_EXPOSE_HEADERS, "X-Custom-ID");
+            headers.add("X-Custom-ID", String.valueOf(city.getId()));
+            headers.setContentType(MediaType.APPLICATION_JSON);
+            return ResponseEntity.ok().headers(headers).body(city);
+        } else {
+            return ResponseEntity.badRequest().body("Cannot add city to the DB - operation failed");
         }
-        return "City successfully deleted!";
     }
 
-    @RequestMapping("{id}")
+    @DeleteMapping("/{id}/delete")
+    @ResponseBody
+    public ResponseEntity<String> delete(@PathVariable("id") Long id) {
+        var city = cityDao.findById(id);
+        if (city.isPresent()) {
+            cityDao.deleteCityById(id);
+            return ResponseEntity.ok("City successfully deleted!");
+        } else {
+            return ResponseEntity.notFound().build();
+        }
+    }
+
+    @GetMapping("{id}")
     @ResponseBody
     public ResponseEntity<?> getById(@PathVariable("id") Long id) {
-        String cityId = "";
-        try {
-            return new ResponseEntity(cityDao.findCityById(id), HttpStatus.OK);
-        } catch (Exception ex) {
-            return new ResponseEntity("City not found", HttpStatus.BAD_REQUEST);
+        var city = cityDao.findById(id);
+        if (city.isPresent()) {
+            return ResponseEntity.ok(city.get());
+        } else {
+            return ResponseEntity.notFound().build();
         }
     }
 }
